+++ src/module/type.ts
@@ -1,0 +1,7 @@
+import type { BaseApiClient } from "@twurple/api/lib/client/BaseApiClient";
+import { ChatClient } from "@twurple/chat";
+
+export type Client = {
+  apiClient: BaseApiClient;
+  chatClient: ChatClient;
+};
